Tests for ApolloURLSession and chunk parsing sequence
